+++ packages/jetbrains-plugin/src/main/kotlin/com/mongodb/jbplugin/inspections/AbstractMongoDbInspectionBridge.kt
@@ -6,6 +6,7 @@ import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.database.dataSource.localDataSource
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.util.Key
+import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.*
 import com.intellij.psi.util.CachedValue
 import com.intellij.psi.util.CachedValueProvider
@@ -28,6 +29,11 @@ abstract class AbstractMongoDbInspectionBridge(
             "QueryForDialect${dialect.javaClass.name}",
         )
 
+    /**
+     * Ktlint complains about buildVisitor being longer than 50 lines but because it is just an object implementation
+     * it should be fine to keep it like this to favor readability
+     */
+    @Suppress("TOO_LONG_FUNCTION")
     override fun buildVisitor(
         holder: ProblemsHolder,
         isOnTheFly: Boolean,
@@ -65,21 +71,28 @@ abstract class AbstractMongoDbInspectionBridge(
                         if (fileInExpression == null || fileInExpression.virtualFile == null) {
                             inspection.visitMongoDbQuery(null, holder, cachedValue!!.value, dialect.formatter)
                         } else {
+                            val cachedQuery = cachedValue!!.value
                             val dataSource =
                                 MongoDbVirtualFileDataSourceProvider().getDataSource(
                                     expression.project,
                                     fileInExpression.virtualFile,
                                 )
+
                             inspection.visitMongoDbQuery(
                                 dataSource?.localDataSource,
                                 holder,
-                                cachedValue!!.value,
+                                queryWithCollectionReference(cachedQuery, fileInExpression.virtualFile),
                                 dialect.formatter,
                             )
                         }
                     }
                 }
             }
+
+            private fun queryWithCollectionReference(query: Node<PsiElement>, virtualFile: VirtualFile) =
+                MongoDbVirtualFileDataSourceProvider().getDatabase(
+                    virtualFile,
+                )?.let { query.queryWithOverwrittenDatabase(it) } ?: query
         }
 }