+++ packages/jetbrains-plugin/src/test/kotlin/com/mongodb/jbplugin/autocomplete/JavaDriverCompletionContributorTest.kt
@@ -1,31 +1,19 @@
 package com.mongodb.jbplugin.autocomplete
 
-import com.intellij.database.dataSource.DatabaseConnectionManager
-import com.intellij.database.dataSource.DatabaseConnectionPoint
-import com.intellij.database.dataSource.LocalDataSource
-import com.intellij.database.dataSource.localDataSource
-import com.intellij.database.psi.DbDataSource
-import com.intellij.database.psi.DbPsiFacade
 import com.intellij.database.util.common.containsElements
-import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
 import com.intellij.testFramework.fixtures.CodeInsightTestFixture
-import com.mongodb.jbplugin.accessadapter.datagrip.DataGripBasedReadModelProvider
 import com.mongodb.jbplugin.accessadapter.slice.GetCollectionSchema
 import com.mongodb.jbplugin.accessadapter.slice.ListCollections
 import com.mongodb.jbplugin.accessadapter.slice.ListDatabases
-import com.mongodb.jbplugin.editor.MongoDbVirtualFileDataSourceProvider
 import com.mongodb.jbplugin.fixtures.*
-import com.mongodb.jbplugin.fixtures.mockDataSource
-import com.mongodb.jbplugin.fixtures.mockDatabaseConnection
 import com.mongodb.jbplugin.mql.BsonObject
 import com.mongodb.jbplugin.mql.BsonString
 import com.mongodb.jbplugin.mql.CollectionSchema
 import com.mongodb.jbplugin.mql.Namespace
 import org.junit.jupiter.api.Assertions.*
 import org.mockito.Mockito.`when`
-import org.mockito.Mockito.mock
 import org.mockito.kotlin.any
 import org.mockito.kotlin.eq
 
@@ -59,7 +47,7 @@ public class Repository {
         psiFile: PsiFile,
         fixture: CodeInsightTestFixture,
     ) {
-        val (dataSource, readModelProvider) = setupConnection(fixture, project)
+        val (dataSource, readModelProvider) = fixture.setupConnection()
 
         `when`(readModelProvider.slice(eq(dataSource), any<ListDatabases.Slice>())).thenReturn(
             ListDatabases(
@@ -112,7 +100,7 @@ public class Repository {
         psiFile: PsiFile,
         fixture: CodeInsightTestFixture,
     ) {
-        val (dataSource, readModelProvider) = setupConnection(fixture, project)
+        val (dataSource, readModelProvider) = fixture.setupConnection()
 
         `when`(readModelProvider.slice(eq(dataSource), eq(ListCollections.Slice("myDatabase")))).thenReturn(
             ListCollections(
@@ -160,7 +148,7 @@ public class Repository {
         psiFile: PsiFile,
         fixture: CodeInsightTestFixture,
     ) {
-        val (dataSource, readModelProvider) = setupConnection(fixture, project)
+        val (dataSource, readModelProvider) = fixture.setupConnection()
         val namespace = Namespace("myDatabase", "myCollection")
 
         `when`(readModelProvider.slice(eq(dataSource), eq(GetCollectionSchema.Slice(namespace)))).thenReturn(
@@ -215,7 +203,7 @@ public class Repository {
         psiFile: PsiFile,
         fixture: CodeInsightTestFixture,
     ) {
-        val (dataSource, readModelProvider) = setupConnection(fixture, project)
+        val (dataSource, readModelProvider) = fixture.setupConnection()
         val namespace = Namespace("myDatabase", "myCollection")
 
         `when`(readModelProvider.slice(eq(dataSource), eq(GetCollectionSchema.Slice(namespace)))).thenReturn(
@@ -270,7 +258,7 @@ public class Repository {
         psiFile: PsiFile,
         fixture: CodeInsightTestFixture,
     ) {
-        val (dataSource, readModelProvider) = setupConnection(fixture, project)
+        val (dataSource, readModelProvider) = fixture.setupConnection()
         val namespace = Namespace("myDatabase", "myCollection")
 
         `when`(readModelProvider.slice(eq(dataSource), eq(GetCollectionSchema.Slice(namespace)))).thenReturn(
@@ -293,39 +281,5 @@ public class Repository {
                 it.lookupString == "myField"
             },
         )
-    }
-
-    private fun setupConnection(
-        fixture: CodeInsightTestFixture,
-        project: Project,
-    ): Pair<LocalDataSource, DataGripBasedReadModelProvider> {
-        val dbPsiFacade = mock<DbPsiFacade>()
-        val dbDataSource = mock<DbDataSource>()
-        val dataSource = mockDataSource()
-        val application = ApplicationManager.getApplication()
-        val realConnectionManager = DatabaseConnectionManager.getInstance()
-        val dbConnectionManager =
-            mock<DatabaseConnectionManager>().also { cm ->
-                `when`(cm.build(any(), any())).thenAnswer {
-                    realConnectionManager.build(it.arguments[0] as Project, it.arguments[1] as DatabaseConnectionPoint)
-                }
-            }
-        val connection = mockDatabaseConnection(dataSource)
-        val readModelProvider = mock<DataGripBasedReadModelProvider>()
-
-        `when`(dbDataSource.localDataSource).thenReturn(dataSource)
-        `when`(dbPsiFacade.findDataSource(any())).thenReturn(dbDataSource)
-        `when`(dbConnectionManager.activeConnections).thenReturn(listOf(connection))
-
-        fixture.file.virtualFile.putUserData(
-            MongoDbVirtualFileDataSourceProvider.Keys.attachedDataSource,
-            dataSource,
-        )
-
-        application.withMockedService(dbConnectionManager)
-        project.withMockedService(readModelProvider)
-        project.withMockedService(dbPsiFacade)
-
-        return Pair(dataSource, readModelProvider)
     }
 }